consumer/requirements.txt
confluent_kafka
pymongo
PyYAML>=5.1.2
fastavro
requests